pyproject.toml
[tool.poetry]
name = "qmatic"
version = "1.0.0"
description = "QPoker official ERC20 token"
authors = ["Johnnie M Martin <Johnnie@qpoker.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.12"
pydantic = "^1.10.4"

[tool.poetry.group.dev.dependencies]
eth-brownie = "^1.19.3"
slither-analyzer = "^0.9.2"
watchfiles = "^0.18.1"
mypy = "^1.0.0"

[tool.mypy]
plugins = [
    "pydantic.mypy"
    ]
# follow_imports = "silent"
# strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
# disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
ignore_missing_imports = true
disallow_untyped_defs = true
namespace_packages = true # opentelemetry

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
